First pass at having preview functionality in instrument builder.
Calls module on server side, so it depends on the PHP part of Loris
being installed. The benefit of this is that we use the same wrappers
and are sure we'll get the same HTML back, doing it in pure javascript
would only be able to guess at the HTML that a server might generate.