Check if the tree element exists before calling a function on it.
When the user doesn't have permission to visit the configuration module, the javascript
breaks since it is calling a function on an element that isn't in the DOM. This commit
first checks if the element exists before calling the function on it.

This was reported because it was breaking other javascript components of the page,
namely the dropdown menues.