Remove duplicate Authorization header causing invalid HTTP request
The UpdateCouchDB user function was adding the Authorization header
to the HTTP request. The postURL function that it called was also
adding the header. This was causing the header to be duplicated in
the HTTP request which was causing the CouchDB user syncing to fail.

Removing it from the updateCouchDB function so that it's only added
in the postURL function should fix this problem.