Reset candidate list filters on clicking "Scan_Done" link
The Scan_done link did not previously reset the filters for the imaging
browser, so if a filter had previously been set it would still be
applied and result in the Imaging Browser showing no results returned.
This adds a reset="true" flag to the HTTP request so that the proper
filter will be applied.