Rebuilding the certification edit event page.
This page will turn into more of a training module, where potential examiners can be trained to administed tests through a series of readings/videos/quizzes. This commit deletes old features from the template, and creates a dropdown where users can select an instrument to be trained on. There is a skeleton structure for future aspects of the module in place as well.