Update GUI for the training module.
Added a "home page" to the training module which allows the examiner
to see which instruments they have already been trained on, which
still require certification and which do not have training. They can
now review old content. Another feature is better quiz validation,
correct dates being inserted into the database for certification,
and modal dialogs that give notifications to the user after the
quiz.