The PHP Database wrapper did not throw exceptions on invalid deletes.
The Database.class.inc file was not throwing exceptions when DELETE
statements failed, only inserts and selects. Instead, DELETE was
silently failing.

This adds an exception for consistency, and also includes the MySQL
error in the exception thrown so that we can see the reason for
statements failing in Travis.