Run tests in PHP 5.4 and 5.6
PHP 5.4 is the earliest version of PHP supported by Travis. Since
we claim to support 5.3, that's the closest we can get to running
our testing suite in PHP 5.3.

PHP 5.6 is the most recent version of PHP. Since our tests should
also pass there, we run our tests with it.

We make the assumption that if it works with 5.4 and 5.6, it should
work with versions in between and don't explicitly test it.