Add a setFakeTableData function to the Database wrapper
This function fakes the data in a table for testing purposes.

It replaces the existing table with a temporary table for the
(testing) session, and then inserts the data passed into the
temporary table. This should allow people to write more robust
data dependant unit tests that depend on the data in the database
without mocking every single query that needs to be used in
that test.