Candidate List: Make DoB date element
The Date of Birth filter is a text element instead of an HTML5 date
element. It tries to use a jQuery plugin to allow a popup, but because
the jQuery plugin doesn't match the new Loris style properly the element
is unreadable. This gets rid of the jQuery plugin and uses a date
element instead of a text field.