Add check for useProjects flag in Radiological Review Module
The radiological review module included a "Project" column and filter,
without bothering to check if useProjects was enabled for the Loris
instance. This adds a check for the config variable. Since the "Project"
filter was right in the middle of the other filters, simply removing
it caused a big hole in the UI when useProjects was turned off, so it's
swapped with the Exclusionary Status filter so that it looks good
whether it's turned on or off.