Check for an equivalent active session when a cancelled session is found.
The CouchDB_Confirm_Integrity script which deletes invalid documents
that got inserted into CouchDB (ie. things that have been deleted from
the database) to ensure consistency between MySQL and CouchDB was
deleting cancelled Visit labels. However, it's possible that the visit
was cancelled and replaced by a new session with the same Visit label,
and this caused the new session to be incorrectly deleted from CouchDB.

This adds a check to ensure that it only deletes cancelled session if
there is no active equivalent.