Make "loris" valid project
If useProjects is set to false, requests are supposed to go through
a default project name of "loris", but the API was not checking the
useProjects flag and hence, claiming that "loris" was an invalid project
when it shouldn't.

This makes the "loris" ProjectID "0" so that all the validation still
works, but we're able to detect if the API should be returning all
data.