Wait for C3 to be loaded before running dashboard JS
The javascript for the dashboard.js gets run immediately
upon loading. However, at the point that the test_name.js
is run the C3 library used by it has not been loaded yet.

This puts it in a jQuery document.ready() handler to ensure
that the C3 library is loaded before the JS is run.