Fix treatment of "All" option in filter tables
The "All" option was not explicitly setting the value of the
dropdown to be empty when constructing a dropdown in JSX, and
as a result the form was submitting the value of the option
(that is, the string "All"), which was causing filters to return
no results since no CenterID matched "All".

This makes it a little smarter about adding the "All" option to
the start of the dropdown and explicitly sets the value to the
empty string.