Revert Reliability to HTML_QuickForm
Override the default page form to use HTML_QuickForm instead of LorisForm.
Reliability instruments aren't yet ready to use LorisForm, as they depend
on a type of QuickForm renderer that is not implemented by LorisForm.
This should be removed once that is fixed, but for now (like
instruments) reliability testing mostly depends on HTML_QuickForm and
the instruments that exist need to be overhauled in order to remove this
dependency.