Remove breadcrumb from error page
The integration tests mostly check for the breadcrumb to
ensure the right page loaded. Since the breadcrumb is
displayed on error pages, this means that the integration
tests pass even if all the module does it throw an
exception.

This removes the breadcrumb if there is an error message,
in order to get more use out of Travis..