Fixed bug where search keyword crashes other menus
The base NDB_Menu_Filter class maintains the filters
as you go between pages. This means that if you set
a search key, and then go to a page which doesn't have
any searchKeywords fields defined, the SQL generated
becomes invalid.

This updates the code so that the filter checks if
$this->searchKeyword is set and non-empty (the field
defined by the form) instead of $this->searchKey['keyword']
(the field with the user input)