Moved the Acknowledgement tab under Tools
This one module has a top level menu item to itself, and since the
name was so long it was causing the menu to wrap on lower screen
resolutions.

This moves the Acknowledgements item to the "Tools" menu and
gets rid of the top level menu, so that it fits on tablets and
lower resolution screens.