Removed references to Password_md5, now that everyone is on PHP 5.5+
The minimum version requirement for LORIS 16.0 was 5.6. This means that,
by 17.0, everyone should have been running on it for a while, and the old
Password_md5 column which was only maintained for compatability with (very) old
versions of PHP that don't have access to the PHP password hashing API is
no longer needed.

This removes the column and code which was only in place as a workaround,
greatly simplifying the logic (and improving the security.)