Added ability to serve markdown help content from module help/ directory
This should be significantly easier to maintain and review than storing
the help content in the database where everything requires an SQL patch
to update. It's also significantly faster.

This is only available for "new" style modules, and the dicom_archive
(which previously had no help at all) is used as an example.