[Core] Added Module base class and improved structure of modules (#2166)
* Added Module base class and improved structure of modules

This adds a "Module" base class for modules to implement, which can be
used for adding information or functions on a module, rather than a per
page basis. The previous design only used the test_name as a hack to determine
which module a page should load from but didn't have any real concept of the
design of "Modules", and no obvious place to put functions like "should this
module be available to anonymous users?"

The new design adds a "LoadPage($pagename)" which returns an NDB_Page instance
in a base LORIS Module class.

Instead of relying on different loadX types for each page type which currently
get delegated to based on confusing rules in NDB_Caller, the new module
structure simply uses autoloading to load a class of the name of the subpage
being loaded, and lets subtype polymorphism take care of the rest. As a result,
pages are easier to find and reason about (they're in the file named
subpagename.class.inc instead of all being jumbled together in a single
NDB_Form_modulename file.)

Individual modules now use a PHP namespace of LORIS\$modulename and files
get autoloaded from the module's php/ directory automatically.

The dicom_archive module is updated as an example of the new module structure.
Other modules should be updated during the development of LORIS 17.x so that we
can remove the legacy code in 18.0.

* Ran PHPCS on dicom archive tests

* Lowercased filesystem names for autoloading

* Fixed bug from case insensitivity on Mac