[Core/Dicom_Archive] Markdown File Help Content (#2196)
This adds the help content from markdown format files in the module `help/` directory.

This should be significantly easier to maintain and review than storing the help content in the database where everything requires an SQL patch to update. It's also faster.

This is only available for "new" style modules, and the dicom_archive (which previously had no help at all) is used as an example.

Only a subset of markdown is supported. Namely:

- Paragraphs (by separating text with an empty line)
- Headers (by prefixing a paragraph with up to six `#` characters. Note that unlike real markdown, there must be an empty line before the first paragraph of the section)
- Bold and italics with `*` or `**`
- Markdown style links

(The new React "Markdown" component can be used to render markdown other places in LORIS, if required.)