LORIS Web Based DB Install (#2010)
This removes all parts of sourcing the schema and updating the config from the
install.sh script and moves it into a friendlier web based installer, where more
error checking and validation can be easily done. Since it's web based, the web
server and PHP need to already be installed and run (as does composer, since it
uses smarty) and the install.sh script can not yet be completely removed.

Being web based, we also have more information that can be deduced instead of
prompted for, such as the hostname/URL (we can just check the hostname that the
user is accessing the install page from) or the host for the MySQL username
(which we can assume is the same as the admin MySQL user creating the database,
since the install script is connecting from the same place as LORIS.)

The API secret token is also now randomly generated when installing the
database, since it doesn't matter what it is as long as the server is always
using the same key to generate and authenticate tokens, and this is easier to
do in PHP than it was in a shell script.