Save instrument data in JSON as well as in tables
This adds support for LORIS to save instrument data into a JSON
formatted column in the flag table, instead of requiring a table
for each instrument.

For now, it only adds the ability to save data and doesn't use it,
but it adds the possibility of cleaning up our schema and requiring
less MySQL permissions to use in the future.