Added the ability for pages to bypass the smarty module.
This adds a "skiptemplate" flag on a page level, which hints at the renderer
that it should not try and load any smarty templates.

This is mostly intended for pages that do their rendering from javascript,
and avoids the need for an empty smarty template just to avoid an exception
because smarty can't open the template file.