[Request Account] Removed exploit to add centers to existing users as non-admin (#2782)
The request_account page was inserting the requested sites to the user, whether or not the account already exists. This resulted in the ability of a user to add sites to arbitrary existing accounts (including their own), by simply requesting it on the request-account page.

This moves the site insertion logic to inside of the check for whether or not the site exists to remove this exploit.