Add the minimal config for phan which LORIS passes
Add the minimal config for LORIS which lets the LORIS
codebase pass with the phan static analysis tool, so that
type errors can be fixed one class of error at a time. See:
https://github.com/etsy/phan/wiki/Incrementally-Strengthening-Analysis

Unfortunately, this means that most classes of errors are
currently ignored, however phan does still do enough analysis to
have picked up on the references to undefined classes in PR#2802.

Ideally, phan would be run as part of the CI, but the requirements
for phan are PHP 7.1 with the AST extension installed, and adding
it to composer.json would mean that Travis wouldn't pass on either
7.0 or 5.6 because of unmet requirements (which aren't actually
requirements for LORIS), so until the minimum requirements for LORIS
are PHP 7.1 this will need to be run manually by developers on occasion.