Fix syntax errors in API (#2802)
This fixes two errors in the API code. BrainBrowser was missing
a semicolon for a statement, and instrument data was trying to
catch "Exception" so that it can return an error if anything
goes wrong, but since the code lives in a namespace it needs
to catch Exception from the root namespace, not it's own
(undefined) Exception type.