[Core] Added safe(r) NDB_Page wrapper around $tpl_data
Add a getter and setter for tpl_data functions, in order to have a place to
sanitize variables for XSS (or other) attacks.

This would ideally be a private to enforce using the getter and setter, but
there are too many modules that directly access it, so such a change would
need to be for 18.0. For now, the infrastructure is put in place to let
people use the getter and setter methods without any defense against directly
mutating the array.

This change only updates the page type base classes in php/libraries to use the
setter and getter, not the modules.