[API] Add check for cryptographic strength of JWT keys (#2865)
Previous LORIS upgrade scripts may have resulted in a weak JWT preshared key being stored in the configuration tables, which was supposed to be a temporary, unused key which is replaced by a randomly generated password during the LORIS install process.

This implements JWT key strength requirements in the API, and refuses to issue tokens for keys that aren't at least 20 characters long, and don't have the special character / number / letter requirements that are in place for user passwords. As a result, the not-so-temporary-afterall key is now blacklisted.