Set smarty's autoescape template variable option
This sets the smarty escape_html option to true by default in the
Smarty_hook class, as well as adding a helper to change the value.

Instruments currently disable autoescaping, because a number of instruments
use hacks for layout which would break if this was changed without notice.

The override for instruments should likely be changed in LORIS 18.0, but
otherwise this should help minimize the risk of XSS attacks in LORIS modules.