[API] candidate creation (#2968)
This makes the EDC field optional if useEDC configSetting is false. It makes the createNew function of the Candidate class throw an exception if useEDC is true and no EDC is provided.

This also associates the candidate to the project providing a valid project name. It update the candidate after its creation.

Note that the candidate is still created if the project name is invalid.