[Core] Reset password not working (#3027)
Fix bug where lost password link was not working.

lost_password calls User::singleton(username) to retrieve user information, which is blocked because NDB_Client calls User::singleton first (using the unix username). This is because the line $client->makeCommandLine() forces NDB_Client to act as if the user is not using a web client.