[Statistics] Removing reliability statistics (#3009)
The reliability statistics tab has too many project specific queries. This removes the tab from the default install but leaves the code in place, so that projects which are currently making use of it can still do so, but new projects don't think the tab is there but not working.