Fix empty result in getMaxThreadStatus for NDB_BVL_Feedback
The NDB_BVL_Feedback class incorrectly had a isset check instead of an !empty
check to determine if an SQL query returned any results. isset would never
be false, because the variable is set directly above the check.

This updates it to be !empty, so that "null" is properly returned when
there is no feedback.