Make framework use traversables instead of arrays
Iterators are able to be lazily evaluated, so converting the
framework to use Traversables rather than arrays allows a more
memory-efficient filtering, as only the rows that are not-filtered
need to be kept in memory.

A DBRowProvisioner is added to take advantage of this, by using
returning a prepared PDO statement, rather than the result of
fetchAll().