[candidate_list] Reactify Open Profile box (#3070)
This updates the "Open Profile" box on the candidate_list page for data entry to use React instead of custom javascript and moves some javascript from the smarty template into a proper javascript file. The data table filters still need to be reactified to get rid of smarty for this module.

The functionality should be unchanged, but the code is more readable and shares more of the UI framework with the rest of LORIS for consistency and maintainability.