[Core] Fixing NDB_Client command line user instantiation (#3212)
Correctly calling User singleton with the accurate username

This came up because when fix_timepoint_date_problems.php calls User::singleton, it passes the current unix user as an argument but it's pretty useless since the static instance is already being initialized at the beginning. Plus the code is incorrectly using the "owner of the file" as the argument, not the user running the script. 

This also fixes the inaccurate error message produced by fix_timepoint_date_problems.php.