[MRI Violations] Fix Site Filter (#3222)
Site filtering was working off of keyword logic i.e., checking that the filtered value is at least a substring of the filtered field. For example, filtering for violated scans from CenterID 2, the menu will get populated with violated scans from CenterID 12, 20, 52, etc.

This fixes it to only check for exact equality matches.