Remove final_radiological_review tables
The module itself was previously removed, but the SQL tables
were not deleted from the SQL schema at the same time.

There is no patch to go with this, because projects who moved the
module to project/modules wouldn't want an upgrade to drop the table
and lose all their data.