[user_accounts] Examiner validation missing 0 bug (#3242)
This modifies a regex to match with zeros where it should have already been the case. Without this PR, on the edit_user page, when a checkbox for examiner at: is associated with a site with a 0 in its ID, the validation will not match it and thus throw an error.