Remove version file
The VERSION file is not properly maintained, and is more confusing
than enlightening for users. Even if it were updated more consistently,
it prevents merging between the bugfix/minor/major branches.

This removes the file so that users aren't misinformed. It's not
maintainable in the current state of LORIS unless there's a way to
automate the updating of it..