[Training] Remove module from LORIS
According to PR#3193, the training module as it exists in LORIS
is not currently used by anyone. This results in an excessive amount
of unnecessary maintainance costs.

Given that the above PR has been open for nearly a month and no one
has been willing to take the time to either review or test the pull
request, while every other module has been updated, maintaining the
training module in LORIS does not appear to be a feasible proposition.

As a result, this removes the module.