[Core] Remove _setupPage function (#3280)
The modularization of LORIS modules had an unfortunate side-effect of re-ordering some of the setup logic, since most modules overrode the _setupPage() function to ensure that LorisForm was instantiated before setting template variables (which may get overridden after _setupPage is called if they did it another way). This also had the annoying side-effect or requiring all 5 parameters just to pass to the parent call.

Since the _setupPage function in the base NDB_Page class was used as a pseudo-constructor, this moves it to a real constructor. The result is that setup() can be called directly and _setupPage can be removed.

This makes it possible to ensure that hasAccess is checked before the setup() function call (some of the hasAccess functions require the identifier attribute which was previously set in the base setupPage, which is now in the constructor.) and prevent the possibility of leaking any data on the "You do not have access page" if setupPage did something which generated a warning, since it's now never called at all if the user doesn't have access.

It also removes the ambiguity of the difference between setup and setupPage.