Publicly accessible module support (#2608)
This adds the ability to create modules which are accessible to the public, and converts the login/password-reset/request-account/etc features into a "login" module, both as an example and because there needs to be a default module to use when not logged in.

Modules must have a Module class in the module namespace, and can be converted to be publicly accessible by overriding the "isPublicModule" function to return true. This causes hasAccess to be bypassed, but modules are otherwise written as before.