Fix router errors in Travis
Fix a bug where Travis is not able to find the htdocs/router.php file,
because it's looking for htdocs/htdocs/router.php. I'm not sure how or
why this started breaking consistently across all versions of PHP being
run by Travis, but it seems to be the case. This takes out the htdocs/
portion of the path used as a router since it seems to be duplicated by
the -t (docroot) of the PHP built in web server.