[Core] Readd password expiry functionality
This re-adds the password expiry feature into loris, which was
accidentally removed in the login module refactoring. (The password
expiry feature will likely be removed in the future, but right
now too many other features such as "Forgot Your Password" depend on it.

This is largely based on the old code, but put into a class in the
login module.