[Instrument] Fix "Delete Instrument Data" feature
The feature that deletes all data from a given LORIS instrument
was broken, because LORIS was incorrectly calling "Save" on
the instrument (which didn't have the correct data posted) rather
than save on the control panel.

This adds a check to ensure that the incorrect "save" isn't
called when using the "Delete instrument data" feature.