[SQL] changes for the MRI side.  (#3327)
Dropped tarchive_find_new_uploads table as it is no longer used in LORIS. However, the drop statement is commented in case older projects have data in that table. Make sure to check that table is empty before uncommenting it.

Added a foreign key constraint to the two following tables so that they are linked to the tarchive table using the TarchiveID field. Affected tables are:

MRICandidateErrors
mri_violations_log

Note: in order to add the foreign key constraints, the creation of the tarchive tables had to be moved before the creation of the mri violation tables.