[Instruments] Remove reference to deprecated PHP function "create_function" (#3341)
This removes a reference to the PHP function "create_function", which was
deprecated in PHP 7.2 in XIN rule evaluation. (create_function was essentially
an `eval`).

Unfortunately, it was used in XIN rules to create a function which can take
an arbitrary operator for comparison, but there is no way to know which operators
instruments are actually using. This adds the known operators after talking
to a couple projects, but throws an exception requesting a bug report for
anyone who was not covered by these cases.