[Statistics] Fix DD Entry statistics links / Load statistics dd site pages (#3368)
This fixes Redmine bug #13699. The loadPage function was returning an NDB_Page subclass with a null Module parameter that is called on in the display method in NDB_Menu.

This  fixes that by setting the Module parameter as $this->Module. (This fix is in the base class, but it didn't affect pages that were overriding setup and not calling their parent.)

See also: https://redmine.cbrain.mcgill.ca/issues/13699