Default config phantom regex (#3443)
This will default the config settings LegoPhantomRegex and LivingPhantomRegex to /lego_phantom/i and /living_phantom/i instead of /./i that allows pretty much everything to be displayed.

This change will avoid project without phantom datasets to update the patientName regex and have the flag "HIDDEN - INVALID" working when the patient name does not match the patientName regex. Because of the way the code is written, it displays "HIDDEN - INVALID" only if patient name does not match PatientNameRegex AND LegoPhantomRegex AND LivingPhantomRegex. If a project does not have phantom datasets, they would not update the LegoPhantomRegex or LivingPhantomRegex which until now defaulted to /./i. Therefore, updating the PatientNameRegex was not enough to hide patient name.