Tool script to run before running MRI patch from PR #3327 (#3438)
Adds a tool script that needs to be run before running the MRI patch sent in PR #3327 or before upgrading to LORIS 19.0

This script will look for orphan entries of the following pairs of table => foreign key:

    MRICandidateErrors => TarchiveID
    mri_violations_log => TarchiveID
    files => TarchiveSource
    files_qcstatus => FileID
    mri_upload => TarchiveID
    mri_upload => SessionID
    mri_protocol_checks => ScanType
    tarchive => SessionID

Take the example of table files_qcstatus that links to the files table using the FileID foreign key. If there is an entry for FileID 3 in files_qcstatus but not in the files table, this SQL patch from PR #3327 will fail creating the FileID foreign key due to this entry with no corresponding FileID in the table of origin. This applies to all pairs of table => foreign key listed above.

Therefore, before running the SQL patch that will create the missing foreign keys on the MRI side, we need to make that no orphan entries can be found in the tables listed above. This script takes care of this. It will:

    fetch orphan entries in those tables
    print orphan entries in the terminal
    ask whether the user want to delete those entries (after proper backup) or update their foreign key to be NULL (the update becomes interesting when dealing with files_qcstatus table as the correct FileID can be repopulated based on the SeriesUID and EchoTime of the file)
    depending on the user's answer ('y' or 'n')

        if 'y', backup the orphan entries using mysqldump into files backup_release_19-0_upgrade.sql in the current directory
        if 'n', update the foreign key value to NULL for the orphan entries