[Upgrade Script] Prompt for MySQL username and password (#3473)
This modifies the way the mysqldump is called in cleanup_mri_tables_for_19-0_release.php.

Instead of using the username and password from the config.xml file, it prompts the user running the script to enter the MySQL username and password to use for mysqldump. The user in the config file does not always have sufficient permissions.