Add page decoration middleware
This adds PSR15 based page decoration middleware for both
anonymous, and logged in users. The middleware takes an undecorated
request and adds all the "chrome" (headers, footers, etc) that
are required to make the requests look consistent and will be
required by an upcoming PSR15-based router pull request.

This is an attempt to split it into smaller, more manageable
chunks. The whole PSR15-based router can be seen in my
PSR7RouterZend branch, for context, but this code should
be self-contained.