[Core-Database] added 2 new functions to help avoid code duplication (#3428)
This adds 2 new functions to the database.class.inc that build on the pselect() and pselectCol() in order to return an array indexed by a selected primary key. These functions are meant to replace the multitude of places in the code where the pselect() and pselectCol() result is ran through a foreach loop in order to extract and re-index the array using its primary keys.

The necessary Exceptions are thrown to avoid data being overwritten in the built array and misuse of the function in general.