Add page decoration middleware (#3477)
This adds PSR15 based page decoration middleware for both
anonymous, and logged in users. The middleware takes an undecorated
request and adds all the "chrome" (headers, footers, etc) that
are required to make the requests look consistent and will be
required by an upcoming PSR15-based router.

This is an attempt to split the router PR into smaller, more manageable
chunks. Note that it implements the handler for NDB_Page, but
isn't currently used.

This also begins using the directory "src/" as a location for properly
namespaced PSR4-based autoloaded classes rather than the
current php/libraries jumble of code used in LORIS.