New data framework interfaces (#3502)
This defines the interfaces for a new LORIS data framework, which is
intended to provide a better separation of concerns than the current 
page framework in which display, data retrieval, data conversion 
(ie. to JSON) and sometimes other concerns are mixed into a single "page"
object which often leaks its abstractions (ie most of the "NDB_Menu_Filter"
class simply builds an SQL query.) The framework
lives in a LORIS\Data namespace, and as the name implies, is only
concerned with managing data, not with displaying it.

The new LORIS data framework is concerned with the retrieval, filtering,
and mapping of data. Unlike the existing "NDB_Page framework" (for lack of
a better name), the new data framework takes more inspiration for its design from
functional than object oriented programming and makes the assumption
that the data to be sent to the user is a pure function of the data itself, and
the user trying to access the data. The interfaces it defines are implemented 
in terms of immutable data structures.

The end result is (hopefully) a cleaner, more maintainable and
debuggable data access framework which makes it easier to implement
features such as subproject or site based permissions consistently across
LORIS modules.

The framework itself takes a fairly high level view of what "data" is
in its interfaces and leaves the specifics of internal representations
to its implementors (ie. module writers) who are closer to the data
itself.

In rough terms, \LORIS\Data\Provisioners get data from somewhere
(such as the LORIS SQL or CouchDB databases) and return \LORIS\Data\DataInstances.
\LORIS\Data\Filters filter the data (ie. for things like permission
checks) before sending it to the user, and \LORIS\Data\Mappers map the data
(ie. for things like anonyization.) See the documentation in the code
for more details.
The dicom_archive is updated as an example/demonstration of how to use this
framework.