Adding description to ImagingFileTypes table (#3464)
This adds a description column to the ImagingFileTypes table.

The ImagingFileType table stores all file type related to imaging and is linked to the FileType column of thefiles table. In the near future, additional file types will be added and some of them might seem obscure so having the capability to add a description for the file type will come in handy.