PSR15 Base Router (#3480)
This builds on #3477 to add the classes which implement (but not yet use) a PSR15 request handler based router for LORIS, as well as adding authentication middleware. 

A PrefixRouter is a type of Router which, based on the prefix of the request, strips off the prefix and delegates to another router.

A BaseRouter serves the root of a LORIS install and, effectively, replaces .htaccess. It annotates the PSR7 request with common properties like the user accessing LORIS and the BaseURL being served, but otherwise acts like a PrefixRouter to strip part of the URL and delegate to another router.

A ModuleRouter is a type of router which adds authentication middleware based on whether or not the module being loaded is public, and delegate's to the module's handler. This is done outside of the module itself's handler to ensure the authentication middleware can't be accidentally removed by a module without setting the public flag.

A ModuleFileRouter is a type of PrefixRouter which handles serving the /css, /js, and /static files of a module off the filesystem.