Entrypoint for new router (#3513)
This adds the index.php entrypoint for the new PSR15 based LORIS router, building on #3480. The router mostly defers to the changes in #3480, but also adds the handler for Modules, and a Content-Length middleware which adds an HTTP Content-Length header to any response.

This doesn't update any existing modules to use it, as some modules that use PHP super-globals will require changes. (The next PR in this series will finally switch over to the new router, and update any existing modules that require changes.)