Move localizer exclusion setting into the Config module (#3449)
Until now, the exclusion of localizers were hardcoded in the tarchiveLoader in the variable $exclude.

This PR allows the possibility to configure the sequences that a user might want to exclude from the dcm2mnc command (localizers, scouts and other). In the "Imaging Pipeline" section of the Config module, it is now possible to specify the different series descriptions to be excluded from the insertion pipeline. Two default example have been set in LORIS: is localizer and scout (see screenshot below)

The imaging script will then used the values stored in the Config modules for the excluded_series_description config setting and create a regular expression out of those values to use afterwards when calling dcm2mnc via the find command which will exclude the files matching the created regular expression pattern.

See also:
Related PR on the LORIS-MRI side: aces/Loris-MRI#264