Remove non-prepared variant of database select wrappers
The non-prepared variant of the database select wrappers have been
deprecated for a long time, but left around because of the need of
projects to update any instruments they might have that uses them.

This finally removes the code, now that users have had time to
update any calls to them and all references in the LORIS codebase
have been removed.