[CouchDB] Adding support for multiple databases on different servers (#3546)
This builds on #2795 to extend the functionality to be able to support multiple databases on different servers. This is backwards compatible as it will revert to the DQT credentials if none are passed. This throws deprecation warnings and exceptions when parameters are not properly passed

This is removing the "auto-load" of credentials from the libraries and explicitely passing them when instantiating a couchDB instance.