[Doc Repo] Unfixed line elements so that they scroll with the frame (#3610)
This address a minor bug in the front end of the Doc repo filter table: 

It unfixes the horizontal lines coming before each file name or category in the tree structure of document repository so that they scroll with the frame. The root of the issue was the use of the pseudo element ::before and the position: absolute. To format a line before each filename and category, the pseudo element is removed and div elements are used instead.

See: Redmine #11267