Fix signature compatibility errors detected by phan
This increases the coverage of phan to detect incompatible
signatures between parent and child classes. (These are primarily
caused by invalid doc comments, where phan extracts the variable
types from, or previous ambiguity between how to denote something
that returns no value. phan insists on "void", while the codebase
had a mixture of "void", "none", or "null".)

When there's disagreement between a child and parent class about
what the return value should be, this change sides with the parent
(which may not always be correct, but can be fixed in a future
change if any errors are detected.)