[JSX] Fixed check for select elements with multiple attribute  (#3406)
This modifies the FilterForm & StaticDataTable components to handle SelectElements where the multiple attribute is selected. Previously, only strings and numbers were handled in the StaticDataTable, and now the 'object' type is being checked for. Additionally, queryString.js now handles an array of values being passed for a single filter element.